+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/addresselement/InputAddressScreen.kt
@@ -18,11 +18,13 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.stripe.android.common.ui.PrimaryButton
+import com.stripe.android.core.strings.ResolvableString
 import com.stripe.android.core.strings.resolvableString
 import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.paymentsheet.R
 import com.stripe.android.paymentsheet.injection.InputAddressViewModelSubcomponent
 import com.stripe.android.paymentsheet.ui.AddressOptionsAppBar
+import com.stripe.android.paymentsheet.ui.ErrorMessage
 import com.stripe.android.paymentsheet.ui.PaymentElementTheme
 import com.stripe.android.ui.core.FormUI
 import com.stripe.android.uicore.elements.CheckboxElementUI
@@ -45,7 +47,8 @@ internal fun InputAddressScreen(
     onCloseClick: () -> Unit,
     topContent: @Composable ColumnScope.() -> Unit,
     formContent: @Composable ColumnScope.() -> Unit,
-    bottomContent: @Composable ColumnScope.() -> Unit
+    bottomContent: @Composable ColumnScope.() -> Unit,
+    saveError: ResolvableString?,
 ) {
     val focusManager = LocalFocusManager.current
     PaymentElementTheme(appearance = appearance) {
@@ -81,6 +84,12 @@ internal fun InputAddressScreen(
                     topContent()
                     formContent()
                     bottomContent()
+                    saveError?.let {
+                        ErrorMessage(
+                            error = it.resolve(),
+                            modifier = Modifier.padding(top = 2.dp),
+                        )
+                    }
                     PrimaryButton(
                         isEnabled = primaryButtonEnabled,
                         isLoading = primaryButtonLoading,
@@ -125,6 +134,7 @@ internal fun InputAddressScreen(
         R.string.stripe_paymentsheet_address_element_shipping_address
     )
     val formEnabled by viewModel.formEnabled.collectAsState()
+    val saveError by viewModel.saveError.collectAsState()
     val checkboxChecked by viewModel.checkboxChecked.collectAsState()
     val billingSameAsShippingState by viewModel.shippingSameAsBillingState.collectAsState()
 
@@ -190,6 +200,7 @@ internal fun InputAddressScreen(
                     }
                 )
             }
-        }
+        },
+        saveError = saveError,
     )
 }